# openapi/railwaysMySQL-api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8081
  description: Generated server url
paths:
  /api/wagons/{wagonId}:
    get:
      tags:
      - wagons
      summary: Get an existing wagon
      operationId: getSingleWagon
      parameters:
      - name: wagonId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                oneOf:
                - $ref: '#/components/schemas/RestaurantWagon'
                - $ref: '#/components/schemas/PassengerWagon'
                - $ref: '#/components/schemas/BedWagon'
                - $ref: '#/components/schemas/MotorWagon'
        "404":
          description: No wagon to get
    put:
      tags:
      - wagons
      summary: Update an existing wagon
      operationId: updateSingleWagon
      parameters:
      - name: wagonId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WagonDOT'
        required: true
      responses:
        "404":
          description: No wagon to update
        "200":
          description: Wagon successfully updated
          content:
            application/json:
              schema:
                type: object
                oneOf:
                - $ref: '#/components/schemas/RestaurantWagon'
                - $ref: '#/components/schemas/PassengerWagon'
                - $ref: '#/components/schemas/BedWagon'
                - $ref: '#/components/schemas/MotorWagon'
        "400":
          description: Parameters not available for type of wagon in body
    delete:
      tags:
      - wagons
      summary: Delete an existing wagon
      operationId: deleteWagon
      parameters:
      - name: wagonId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: No wagon to delete
        "204":
          description: No content
...
  /api/wagons:
    get:
      tags:
      - wagons
      summary: Get wagons list with optional filters
      operationId: findWagonPage
      parameters:
      - name: bathroom
        in: query
        required: false
        schema:
          type: boolean
      - name: wagonClass
        in: query
        required: false
        schema:
          type: string
          enum:
          - PRIMA
          - SECONDA
          - ECONOMY
      - name: wagonType
        in: query
        required: false
        schema:
          type: string
          enum:
          - MOTOR
          - PASSENGER
          - BED
          - RESTAURANT
      - name: minimumTables
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: kitchenType
        in: query
        required: false
        schema:
          type: string
      - name: fuelType
        in: query
        required: false
        schema:
          type: string
          enum:
          - ELECTRIC
          - DIESEL
      - name: minimumBeds
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: bedType
        in: query
        required: false
        schema:
          type: string
      - name: minimumSeats
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: offset
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: limit
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        "400":
          description: Bad request
    post:
      tags:
      - wagons
      summary: Add a new wagon
      operationId: createWagon
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WagonDOT'
        required: true
      responses:
        "409":
          description: Parameters not available for type of wagon in body
        "201":
          description: Created
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/RestaurantWagon'
                - $ref: '#/components/schemas/PassengerWagon'
                - $ref: '#/components/schemas/BedWagon'
                - $ref: '#/components/schemas/MotorWagon'
...
  /api/trains/{trainId}/wagon/{wagonId}:
    post:
      tags:
      - wagons-trains
      summary: Add or rewrite a new association between wagon and train
      operationId: createOrUpdateTrainWagonAssociation
      parameters:
      - name: trainId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: wagonId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "409":
          description: Already exists an association between ids in path
        "201":
          description: Created
...
components:
  schemas:
    WagonDOT:
      type: object
      properties:
        id:
          type: integer
          format: int32
        trainId:
          type: integer
          format: int32
        wagonClass:
          type: string
          enum:
          - PRIMA
          - SECONDA
          - ECONOMY
        wagonType:
          type: string
          enum:
          - MOTOR
          - PASSENGER
          - BED
          - RESTAURANT
        length:
          type: number
          format: double
        width:
          type: number
          format: double
        height:
          type: number
          format: double
        bathroom:
          type: boolean
        fuelType:
          type: string
          enum:
          - ELECTRIC
          - DIESEL
        kitchenType:
          type: string
        bedType:
          type: string
        nseats:
          type: integer
          format: int32
        ntables:
          type: integer
          format: int32
        nbeds:
          type: integer
          format: int32
    RestaurantWagon:
      type: object
      properties:
        id:
          type: integer
          format: int32
        train:
          type: integer
          format: int32
        wagonClass:
          type: string
          enum:
          - PRIMA
          - SECONDA
          - ECONOMY
        length:
          type: number
          format: double
        width:
          type: number
          format: double
        height:
          type: number
          format: double
        bathroom:
          type: boolean
        wagonType:
          type: string
          enum:
          - MOTOR
          - PASSENGER
          - BED
          - RESTAURANT
        kitchenType:
          type: string
        ntables:
          type: integer
          format: int32
    PassengerWagon:
      type: object
      properties:
        id:
          type: integer
          format: int32
        train:
          type: integer
          format: int32
        wagonClass:
          type: string
          enum:
          - PRIMA
          - SECONDA
          - ECONOMY
        length:
          type: number
          format: double
        width:
          type: number
          format: double
        height:
          type: number
          format: double
        bathroom:
          type: boolean
        wagonType:
          type: string
          enum:
          - MOTOR
          - PASSENGER
          - BED
          - RESTAURANT
        nseats:
          type: integer
          format: int32
    BedWagon:
      type: object
      properties:
        id:
          type: integer
          format: int32
        train:
          type: integer
          format: int32
        wagonClass:
          type: string
          enum:
          - PRIMA
          - SECONDA
          - ECONOMY
        length:
          type: number
          format: double
        width:
          type: number
          format: double
        height:
          type: number
          format: double
        bathroom:
          type: boolean
        wagonType:
          type: string
          enum:
          - MOTOR
          - PASSENGER
          - BED
          - RESTAURANT
        bedType:
          type: string
        nbeds:
          type: integer
          format: int32
    MotorWagon:
      type: object
      properties:
        id:
          type: integer
          format: int32
        train:
          type: integer
          format: int32
        wagonClass:
          type: string
          enum:
          - PRIMA
          - SECONDA
          - ECONOMY
        length:
          type: number
          format: double
        width:
          type: number
          format: double
        height:
          type: number
          format: double
        bathroom:
          type: boolean
        wagonType:
          type: string
          enum:
          - MOTOR
          - PASSENGER
          - BED
          - RESTAURANT
        fuelType:
          type: string
          enum:
          - ELECTRIC
          - DIESEL
...
    Page:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            type: object
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/Pageable'
        empty:
          type: boolean
    Pageable:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/Sort'
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
        unpaged:
          type: boolean
        paged:
          type: boolean
    Sort:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean